refactor(framework): extract EventBus from CraftContext
Moves handler registration (exact + wildcard), payload construction,
dispatch, and the pattern matchers verbatim from context.ts into a new
event-bus.ts; CraftContext.on/once/emit are now 1-line delegations with
identical public signatures. context.ts drops from 1,004 to 816 lines
and eventing becomes a single-responsibility unit. EventBus is not
exported from the package index; the public surface is unchanged.
